+++ oinkbrew/Settings.h
@@ -37,4 +37,5 @@ extern const short LISTENER_PORT;
 
 extern const unsigned long DURATION_MESSAGE;
 
+
 #endif