+++ screens/fonts.h
@@ -28,14 +28,14 @@
 #define __FONTS_H
 
 
-#define FONT_ARIAL_8		0
 #define FONT_ARIAL_10		1 
-#define FONT_ARIAL_12		2
-#define FONT_ARIAL_16		3
+#define FONT_ARIAL_11		2 
+#define FONT_ARIAL_12		3
+#define FONT_ARIAL_16		4
 
 
-extern const D4D_FONT_DESCRIPTOR d4dfnt_Arial_8_desc;
 extern const D4D_FONT_DESCRIPTOR d4dfnt_Arial_10_desc;
+extern const D4D_FONT_DESCRIPTOR d4dfnt_Arial_11_desc;
 extern const D4D_FONT_DESCRIPTOR d4dfnt_Arial_12_desc;